spread logic out to individual files
